Squashed commit of the following:

commit c4541b26a1e895d7674f3a52bdfabef9a25072a0
Author: Maya Dagon <mdagon@cisco.com>
Date:   Wed Jun 2 17:13:12 2021 -0400

    http_inspect: update test mock api

+++ src/service_inspectors/http_inspect/test/http_msg_head_shared_util_test.cc
@@ -37,8 +37,8 @@ using namespace snort;
 // Stubs whose sole purpose is to make the test code link
 long HttpTestManager::print_amount {};
 bool HttpTestManager::print_hex {};
-void FlowData::update_allocations(unsigned long) {}
-void FlowData::update_deallocations(unsigned long) {}
+void FlowData::update_allocations(size_t) {}
+void FlowData::update_deallocations(size_t) {}
 
 // Tests for get_next_code()
 TEST_GROUP(get_next_code)
